+++ internal/agent/ownership_other.go
@@ -9,4 +9,6 @@ import "github.com/zyvorai/netra/internal/models"
 
 func (a *Agent) enrichSocketOwnership(_ []models.TCPHealthStat, _ []models.ProcessMetaStat) {}
 
-func (a *Agent) watchCapChanges(_ []models.ProcessMetaStat) []models.CapChangeEvent { return nil }
+func (a *Agent) watchCapChanges(_ []models.ProcessMetaStat, _ map[uint32]uint64) []models.CapChangeEvent {
+	return nil
+}